+++ agora/service/models.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 from agora import  settings
 from django.db import models
+from django.utils import timezone
 import uuid
 from owner.models import ContactInformation
 from common import helper
@@ -250,6 +251,11 @@ class Resource(models.Model):
             max_length=30,
             default='draft')
 
+    # Datetime fields
+    created_at = models.DateTimeField(auto_now_add=True, null=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    published_at = models.DateTimeField(blank=True, null=True)
+
     def __unicode__(self):
         return str(self.erp_bai_0_id)
 
@@ -293,12 +299,22 @@ class Resource(models.Model):
         return ", ".join(o.erp_bai_0_id for o in self.related_resources.all())
 
     @property
+    def erp_dei_1_required_resources_public(self):
+        return self.required_resources.filter(state='published')
+
+    @property
+    def erp_dei_2_related_resources_public(self):
+        return self.related_resources.filter(state='published')
+
+    @property
     def erp_cli_5_target_users_verbose(self):
         return ", ".join(o.user for o in self.erp_cli_5_target_users.all())
 
     def save(self, *args, **kwargs):
         self.erp_bai_0_id = self.erp_bai_0_id.strip()
         self.erp_bai_1_name = self.erp_bai_1_name.strip()
+        if self.state == 'published' and not self.published_at:
+            self.published_at = timezone.now()
         clean_html_fields(self)
         super(Resource, self).save(*args, **kwargs)