+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3](https://github.com/grnet/agora-sp/compare/v1.1.2...v1.1.3) - 2021-02-03
+
+### Added
+- Add page title configuration option.
+- Add Agora favicon.
+- Add created_at, updated_at fields to Provider, Resource and ContactInformation.
+- Add updated_at field to User.
+
 ## [1.1.2](https://github.com/grnet/agora-sp/compare/v1.1.1...v1.1.2) - 2021-01-13
 
 ### Fixed